requirements.txt
paho-mqtt==1.3.1
py-irsend==1.0.0
